Move type definitions to separate header file.
